pprof_util

Struct StackProfile

pub struct StackProfile {
    pub annotations: Vec<String>,
    pub stacks: Vec<(WeightedStack, Option<usize>)>,
    pub mappings: Vec<Mapping>,
}

A minimal representation of a profile that can be parsed from the jemalloc heap profile.

Fields
annotations: Vec<String>stacks: Vec<(WeightedStack, Option<usize>)>mappings: Vec<Mapping>
Implementations

impl StackProfile

pub fn to_pprof( &self, sample_type: (&str, &str), period_type: (&str, &str), anno_key: Option<String>, ) -> Vec<u8>

Converts the profile into the pprof format.

pprof encodes profiles as gzipped protobuf messages of the Profile message type (see pprof/profile.proto).

pub fn to_flamegraph(&self, opts: &mut FlamegraphOptions<'_>) -> Result<Vec<u8>>

Converts the profile into a flamegraph SVG, using the given options.

impl StackProfile

pub fn push_stack(&mut self, stack: WeightedStack, annotation: Option<&str>)

pub fn push_mapping(&mut self, mapping: Mapping)

pub fn iter(&self) -> StackProfileIter<'_>